+++ tests/ValidParamsTest.cpp
@@ -47,6 +47,17 @@ TEST (ValidParamsTest, params_shouldValidateTextureUnitParams) {
     ASSERT_EQ(0, parser->getExceptions()->size());
 }
 
+TEST (ValidParamsTest, params_shouldValidateSampleParams) {
+    auto *parser = new OgreScriptLSP::Parser();
+    std::string scriptFile = "./examples/params/valid_params_sample_block.material";
+    parser->loadScript(scriptFile);
+
+    parser->parse();
+    OgreScriptLSP::ParamsValidator::getSingleton()->paramsAnalysis(parser);
+
+    ASSERT_EQ(0, parser->getExceptions()->size());
+}
+
 
 void validateChild(OgreScriptLSP::ParamsTree *&tree,
                    const std::vector<std::string> &tokenNames,